Changed the like/disliked output to ints

+++ plugins/rottentomatoes.py
@@ -52,7 +52,7 @@ def rotten_tomatoes(text, bot):
     reviews = review_request.json()
     review_count = reviews['total']
 
-    fresh = critics_score * review_count / 100
+    fresh = int(critics_score * review_count / 100)
     rotten = review_count - fresh
 
     return "\x02{}\x02 - Critics Rating: \x02{}%\x02 ({} liked, {} disliked), " \
